requirements.txt
# Core dependencies
openai==2.16.0
python-dotenv==1.2.1
pillow==12.1.0

# Testing dependencies
pytest==9.0.2
pytest-cov==7.0.0
